+++ release-notes.md
@@ -7,6 +7,7 @@ for [Chrome extension to for Bible.com](README.md).
 ## 2.4.0 (🔜 coming soon)
 
 - [x] Duplicate Slide with Context Menu (right click)
+- [x] Check duplicate content when import settings
 
 ## 2.3.0